requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-marshmallow
marshmallow-sqlalchemy
marshmallow
